requirements.txt
# general
numpy
pandas
matplotlib
python-dotenv

# image manipulation
opencv-python
pillow

# postgres database
psycopg2-binary

# streetview images
streetlevel

# YOLO, tree segmentation/detection model
ultralytics

# for gemini flash model
google-generativeai

# for openai gpt-4o-mini
openai

# for aws
boto3
